Fix issue #5
